Add files via upload
In this project, I have made an interactive navigation bar using HTML5, CSS & JavaScript. Upon landing, users are able to see the nav-bar with 6 menu options. When hovered over them, each menu items display a background respective to my pets with some details and their lives. When the window is resized, the nav-bar changes into a side-bar maintaining the same effects. This is a work in-development and is yet to be completed.